+++ gsa/src/web/pages/tasks/__tests__/detailspage.js
@@ -51,7 +51,10 @@ const config = ScanConfig.fromElement({
   scanner: {name: 'scanner1', type: '0'},
   type: OPENVAS_SCAN_CONFIG_TYPE,
   tasks: {
-    task: [{id: '12345', name: 'foo'}, {id: '678910', name: 'task2'}],
+    task: [
+      {id: '12345', name: 'foo'},
+      {id: '678910', name: 'task2'},
+    ],
   },
 });
 
@@ -1153,7 +1156,7 @@ describe('Task ToolBarIcons tests', () => {
     expect(handleTaskStart).not.toHaveBeenCalled();
     expect(icons[7]).toHaveAttribute(
       'title',
-      'Permission to start Task denied',
+      'Permission to start task denied',
     );
 
     fireEvent.click(icons[8]);