+++ gsa/src/web/pages/tasks/icons/__tests__/starticon.js
@@ -19,6 +19,7 @@ import React from 'react';
 
 import Capabilities from 'gmp/capabilities/capabilities';
 
+import Audit, {AUDIT_STATUS} from 'gmp/models/audit';
 import Task, {TASK_STATUS} from 'gmp/models/task';
 
 import {rendererWith, fireEvent} from 'web/utils/testing';
@@ -72,7 +73,36 @@ describe('Task StartIcon component tests', () => {
     fireEvent.click(element);
 
     expect(clickHandler).not.toHaveBeenCalled();
-    expect(element).toHaveAttribute('title', 'Permission to start Task denied');
+    expect(element).toHaveAttribute('title', 'Permission to start task denied');
+    expect(element).toHaveStyleRule('fill', Theme.inputBorderGray, {
+      modifier: `svg path`,
+    });
+  });
+
+  test('should render in inactive state if wrong command level permissions for audit are given', () => {
+    const caps = new Capabilities(['everything']);
+    const audit = Audit.fromElement({
+      status: AUDIT_STATUS.new,
+      target: {_id: '123'},
+      permissions: {permission: [{name: 'get_task'}]},
+      usage_type: 'audit',
+    });
+    const clickHandler = jest.fn();
+
+    const {render} = rendererWith({capabilities: caps});
+
+    const {element} = render(<StartIcon task={audit} usageType="audit" />);
+
+    expect(caps.mayOp('start_task')).toEqual(true);
+    expect(audit.userCapabilities.mayOp('start_task')).toEqual(false);
+
+    fireEvent.click(element);
+
+    expect(clickHandler).not.toHaveBeenCalled();
+    expect(element).toHaveAttribute(
+      'title',
+      'Permission to start audit denied',
+    );
     expect(element).toHaveStyleRule('fill', Theme.inputBorderGray, {
       modifier: `svg path`,
     });